<fix>[block_storage]: remove get config file from fs id in linux.py
Resolves: ZSTAC-71933
