+++ zstacklib/zstacklib/utils/linux.py
@@ -25,7 +25,6 @@ import json
 import fcntl
 import simplejson
 import xxhash
-import ConfigParser
 
 from inspect import stack
 import xml.etree.ElementTree as etree
@@ -241,23 +240,6 @@ def on_debian_based(distro=None, exclude=[]):
         return innner
     return wrap
 
-def get_config_path_from_fs_id(fs_id):
-    pscli_path = "/usr/local/bin/pscli"
-    if os.path.exists(pscli_path):
-        config_path = "/etc/xstor_{0}.conf".format(fs_id)
-        if os.path.exists(config_path):
-            return config_path
-        default_config_path = "/etc/xstor.conf"
-        if not os.path.exists(default_config_path):
-            raise Exception("no configuration file path is matched, system_id: {0}.".format(fs_id))
-        config = ConfigParser.ConfigParser()
-        config.read(default_config_path)
-        system_id = config.get('xstor', 'system_id', None)
-        if system_id is None or system_id != str(fs_id):
-            raise Exception("no configuration file path is matched, system_id: {0}.".format(fs_id))
-        return default_config_path
-    else:
-        raise Exception("no configuration file path is matched, fs id: {0}.".format(fs_id))
 
 def get_current_timestamp():
     return time.mktime(datetime.datetime.now().timetuple())
